requirements.txt
psycopg2-binary==2.9.9
python-decouple==3.8
telebot==0.0.5
SQLAlchemy==2.0.30
